admin_untact/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-mssql-backend==2.8.1
djangorestframework==3.11.0
idna==3.3
mssql-django==1.0
netaddr==0.8.0
pyodbc==4.0.32
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
django-cors-headers==3.10.0